requirements.txt
streamlit
streamlit-oauth
python-dotenv
pandas
altair
